refactor: seperate some code from Redpacket
